Various pose estimation kernels work with our harness interface and were tested on a g474. Hardware delay was added because software based delay did not work properly and had high energy consumption. Began a matmul ubmark set of kernels that is somewhat unfinished. Tests need to be written and the problem spec MatMulProblem needs to be finished.

+++ src/ento-pose/data_gen.cc
@@ -377,7 +377,7 @@ void display_result(const std::vector<EntoPose::BenchmarkResult<Scalar>> &result
 
 int main(int argc, char* argv[])
 {
-  using Scalar = double;
+  using Scalar = float;
   std::vector<EntoPose::BenchmarkResult<Scalar>> results;
 
   // Argument parsing
